worker_services/emailservice1/requirements.txt
pika==1.3.2
sendgrid==6.11.0
firebase-admin==6.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
